requirements.txt
aenum==3.1.15
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.109.2
frozenlist==1.4.1
fsspec==2024.2.0
gradientai==1.7.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
joblib==1.3.2
llama-index==0.10.5
llama-index-agent-openai==0.1.1
llama-index-core==0.10.5
llama-index-embeddings-gradient==0.1.1
llama-index-embeddings-openai==0.1.1
llama-index-legacy==0.9.48
llama-index-llms-gradient==0.1.1
llama-index-llms-openai==0.1.2
llama-index-multi-modal-llms-openai==0.1.1
llama-index-program-openai==0.1.1
llama-index-question-gen-openai==0.1.1
llama-index-readers-file==0.1.3
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
numpy==1.24.4
openai==1.12.0
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pydantic==1.10.14
PyMuPDF==1.23.22
PyMuPDFb==1.23.22
pypdf==4.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
starlette==0.36.3
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.2.0
uvicorn==0.27.1
wrapt==1.16.0
yarl==1.9.4
